requirements.txt
cryptocompare==0.7.6
matplotlib==3.6.3
psycopg2-binary==2.9.6
pyTelegramBotAPI==4.22.1
Requests==2.32.3
